+++ src/rules/js/__tests__/outOfSyncDeps.test.js
@@ -196,6 +196,87 @@ describe('jsOutOfSyncDeps', () => {
       expect(global.warn).not.toHaveBeenCalled();
       expect(global.fail).toHaveBeenCalled();
     });
+
+    it('should not warn about local (file) dependencies when other deps are in sync', () => {
+      const validDeps = {
+        dependencies: {
+          dep1: '~1.2.3',
+          dep3: '^3.4.5',
+          common: 'file:../common',
+        },
+      };
+      const packageLockJson = {
+        packages: {
+          'node_modules/dep1': {
+            version: '1.2.3',
+          },
+          'node_modules/dep3': {
+            version: '3.4.5',
+          },
+        },
+      };
+      const mockFiles = {
+        'package.json': JSON.stringify(validDeps),
+        'package-lock.json': JSON.stringify(packageLockJson),
+      };
+      fs.setMockFiles(mockFiles);
+
+      jsOutOfSyncDeps();
+
+      expect(global.warn).not.toHaveBeenCalled();
+    });
+
+    it('should not warn about local (file) dependencies when other deps are not in sync (warn only for those)', () => {
+      const packageJson = {
+        dependencies: {
+          dep1: '~1.2.3',
+          dep2: '^2.3.4',
+          dep3: '^3.4.5',
+          dep4: '^4.5.6',
+          common: 'file:../common',
+        },
+      };
+      const packageLockJson = {
+        packages: {
+          'node_modules/dep1': {
+            version: '1.2.4',
+          },
+          'node_modules/dep3': {
+            version: '3.0.5',
+          },
+          'node_modules/dep4': {
+            version: '4.5.6',
+          },
+        },
+      };
+      const mockFiles = {
+        'package.json': JSON.stringify(packageJson),
+        'package-lock.json': JSON.stringify(packageLockJson),
+      };
+      fs.setMockFiles(mockFiles);
+
+      const expectedDep1 = buildDepMessage('dep1', '1.2.3', '1.2.4');
+      const expectedDep2 = buildDepMessage('dep2', '2.3.4', '');
+      const expectedDep3 = buildDepMessage('dep3', '3.4.5', '3.0.5');
+
+      jsOutOfSyncDeps();
+
+      expect(global.warn).toHaveBeenCalledWith(
+        expect.stringMatching(expectedDep1),
+      );
+      expect(global.warn).toHaveBeenCalledWith(
+        expect.stringMatching(expectedDep2),
+      );
+      expect(global.warn).toHaveBeenCalledWith(
+        expect.stringMatching(expectedDep3),
+      );
+      expect(global.warn).not.toHaveBeenCalledWith(
+        expect.stringMatching('dep4'),
+      );
+      expect(global.warn).not.toHaveBeenCalledWith(
+        expect.stringMatching('common'),
+      );
+    });
   });
 
   describe('Old lockfile format', () => {